EC-904 Moved GetEmailAddressForCommentsAndAnswers to Context
