+++ Comments/Models/Context.cs
@@ -1,4 +1,4 @@
-using Comments.Services;
+﻿using Comments.Services;
 using Comments.ViewModels;
 using Microsoft.EntityFrameworkCore;
 using System;
@@ -965,5 +965,21 @@ namespace Comments.Models
 				.Any(c => c.OrganisationUser.OrganisationAuthorisation.OrganisationId.Equals(organisationId));
 
 		}
-	}
+
+        public List<string> GetEmailAddressForCommentsAndAnswers(CommentsAndQuestions commentsAndQuestions)
+        {
+            var commentIds = commentsAndQuestions.Comments.Select(c => c.CommentId).ToList();
+            var answerIds = commentsAndQuestions.Questions.SelectMany(q => q.Answers).Select(a => a.AnswerId).ToList();
+
+
+            var emailAddresses = OrganisationUser
+                .Where(o => Comment.Where(c => commentIds.Contains(c.CommentId)).Select(c => c.OrganisationUserId).Contains(o.OrganisationUserId)
+                         || Answer.Where(a => answerIds.Contains(a.AnswerId)).Select(a => a.OrganisationUserId).Contains(o.OrganisationUserId))
+                .Select(o => o.EmailAddress)
+                .Distinct()
+                .ToList();
+
+            return emailAddresses;
+        }
+    }
 }
